backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
yfinance==0.2.40
anthropic==0.25.0
python-dotenv==1.0.1
